+++ lib/chat_context_menu.dart
@@ -6,3 +6,4 @@ export 'src/ui/chat_context_menu_wrapper.dart';
 export 'src/ui/chat_selectable_text.dart';
 export 'src/model/statement.dart';
 export 'src/model/menu_trigger_mode.dart';
+export 'src/model/menu_animation_style.dart';